+++ .github/workflows/ci.yml
@@ -469,7 +469,7 @@ jobs:
       - uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4
         with:
           fetch-depth: 0
-      - name: Check file sizes (max 500 lines)
+      - name: Check file sizes (max 600 lines)
         # Scoped to files THIS PR changed under packages/studio. Walking the
         # whole tree blamed every unrelated PR for pre-existing offenders.
         # Falls back to a full scan on push events (no base ref available)
@@ -494,10 +494,9 @@ jobs:
           for f in "${files[@]}"; do
             [ -z "$f" ] && continue
             [ -f "$f" ] || continue  # skip files deleted in this PR
-            if grep -qxF "$f" .filesize-allowlist 2>/dev/null; then continue; fi
             lines=$(wc -l < "$f")
-            if [ "$lines" -gt 500 ]; then
-              echo "::error file=$f::$f has $lines lines (max 500)"
+            if [ "$lines" -gt 600 ]; then
+              echo "::error file=$f::$f has $lines lines (max 600)"
               EXIT=1
             fi
           done
